BTD6 dark-render fix: Effects + Minions on every Pro view, hero aura grounding
Closes decode-status item 6d on top of the item-5 Ask parity commit:

- The buff/zone renderers + tier_effect_lines moved from
  btd6_upgrade_detail_service to utils/btd6/effect_lines.py
  (helper-policy: needed by services AND utils -> lives in utils;
  back-compat aliases keep the service's internals and tests stable).
- The shared Pro body (_stat_node_embed) renders two new sections —
  Effects (buffs/zones) and Minions (subtowers w/ headline + lifespan;
  the 9,999,999 sentinel reads 'permanent') — so tower, hero, AND
  paragon Pro views all surface them; whole-bullet truncation with an
  honest drop count keeps every field within Discord's 1024 cap
  (re-validated across all 2,022 embeds).
- The hero grounding emits change-only [btd6_hero_buff] aura lines
  (order-insensitive set comparison — buff order shifts between levels).
- Found by rendering: Striker's Bomb buff carried the dump's misleading
  *Multiplier field names and read as x0.25/x0.05 REDUCTIONS — remapped
  to the *Percentage family (+25% pierce, +5% range) with
  _CUTOVER_TRANSPLANT_SKIP entries so the old committed fields cannot
  ride the transplant back; striker_jones.json regenerated; audit stays
  0 DELTA / 0 SUSPECT.

New tests: embed Effects/Minions + bullet truncation + the permanent
sentinel; hero-aura change-only grounding incl. the fraction fix; the
striker decode test updated. Full CI mirror green; arch 0 errors.
