+++ docs/current-state.md
@@ -6,7 +6,7 @@
 > live GitHub** before trusting it (two same-session reports already
 > contradicted each other across a single merge).
 >
-> **▶ Next action:** **One live queue — [`planning/consolidated-implementation-plan-2026-06-10.md`](planning/consolidated-implementation-plan-2026-06-10.md)** (2026-06-10: the two untapped mapping PRs **#646**/**#647** merged + source-verified, queue-truth drift fixed, the 06-09 gated tail carried item-by-item — its §6 table is the audit trail). **Batch 1** (low-risk runtime truth/clarity — **executed 2026-06-10 in #650**, parallel-lane session; verify merged) **+ Batch 2** (surface-classification completeness invariant — **shipped 2026-06-10 in #651**) both ran as the parallel pair; **Batch 3** service boundaries (RS03/RS06/RS17 — **shipped #652**; RS07 chain-service slice still open) → **Batch 4** Settings **Phase 2** core (declaration coverage + Q-0064 BTD6 rows — **shipped 2026-06-10 in #654**; pointer-classification tail open) → **Batch 5** adaptive **P1C** (Q-0032 — **executed 2026-06-10 in #656**, verify merged) → **next up: Batch 6** Help **projection seam** (Q-0055–Q-0059 overlay after). **Batch 7** mutation hardening is decision-cleared — **Q-0071–Q-0074 all answered 2026-06-10** (recommended options; router §31: workflow-service transactions · workshop boundary first · `setlogchannel` Settings-projected · Admin placement admin-visible); BTD6 `--all` cutover: **#649 merged 2026-06-10; post-cutover VERIFIED + every carry-forward DECODED the same day (#653 wave 1 ∥ PR #655 — verification: dump fidelity byte-identical · 2,022 menu embeds in-limits · AI battery green · fixes for mode-rules dark data / `!btd6 diagnostics` 400 / stamp-rot / path leak; decode: `_CUTOVER_CARRYFORWARD` empty, audit 91 CLEAN / 0 DELTA / 0 SUSPECT, banana economy answerable; next: decode-status ⭐ item 3, the buff/zone tail)**. The Q-0036 denial-copy wiring stays gated on the maintainer's markup of the #632 table. *(The 8-lane scoreboard completed 2026-06-09/10 — record: [`planning/multi-lane-execution-plan-2026-06-09.md`](planning/multi-lane-execution-plan-2026-06-09.md), now `historical`.)*
+> **▶ Next action:** **One live queue — [`planning/consolidated-implementation-plan-2026-06-10.md`](planning/consolidated-implementation-plan-2026-06-10.md)** (2026-06-10: the two untapped mapping PRs **#646**/**#647** merged + source-verified, queue-truth drift fixed, the 06-09 gated tail carried item-by-item — its §6 table is the audit trail). **Batch 1** (low-risk runtime truth/clarity — **executed 2026-06-10 in #650**, parallel-lane session; verify merged) **+ Batch 2** (surface-classification completeness invariant — **shipped 2026-06-10 in #651**) both ran as the parallel pair; **Batch 3** service boundaries (RS03/RS06/RS17 — **shipped #652**; RS07 chain-service slice still open) → **Batch 4** Settings **Phase 2** core (declaration coverage + Q-0064 BTD6 rows — **shipped 2026-06-10 in #654**; pointer-classification tail open) → **Batch 5** adaptive **P1C** (Q-0032 — **executed 2026-06-10 in #656**, verify merged) → **next up: Batch 6** Help **projection seam** (Q-0055–Q-0059 overlay after). **Batch 7** mutation hardening is decision-cleared — **Q-0071–Q-0074 all answered 2026-06-10** (recommended options; router §31: workflow-service transactions · workshop boundary first · `setlogchannel` Settings-projected · Admin placement admin-visible); BTD6 `--all` cutover: **#649 merged 2026-06-10; post-cutover VERIFIED + every carry-forward DECODED the same day (#653 wave 1 ∥ PR #655 — verification: dump fidelity byte-identical · 2,022 menu embeds in-limits · AI battery green · fixes for mode-rules dark data / `!btd6 diagnostics` 400 / stamp-rot / path leak; decode: `_CUTOVER_CARRYFORWARD` empty, audit 91 CLEAN / 0 DELTA / 0 SUSPECT, banana economy answerable; next: decode-status ⭐ item 3, the buff/zone tail)**; **answerability items 5+6d shipped 2026-06-10 in PR #658** (deterministic Ask answers bloons/powers/MK/bosses; Pro views render Effects+Minions; Striker fraction-semantics fix). The Q-0036 denial-copy wiring stays gated on the maintainer's markup of the #632 table. *(The 8-lane scoreboard completed 2026-06-09/10 — record: [`planning/multi-lane-execution-plan-2026-06-09.md`](planning/multi-lane-execution-plan-2026-06-09.md), now `historical`.)*
 >
 > 1. **Mining character platform** — Wave 1 continues. Shipped: foundation **#606**, "The Descent" **#607**, combat gear→deathmatch **#608**, sell/buy market **#609**, Character overview **#610**; the **Workshop + durability keystone and the §6.3 mother-panel live overview were built 2026-06-09 in PR #624** (migration 063 `mining_gear_wear`, `cogs/mining/workshop.py`, Workshop panel + `!repair`/`!craft`/`!quickcraft` — merged, verified live 2026-06-09; tuning confirmed as shipped + duels-wear queued, **Q-0054 answered**). **Next slice — Q-0072 answered (2026-06-10): the workshop-workflow service boundary first** (mapping FIND-RS02), then structures (§7.5) / game-XP (§7.4). Route in: `docs/ideas/mining_exploration_brainstorm.md` §7.7 + the games folio.
 > 2. **Adaptive Setup/Access platform** — Phase 0 complete; Phase 1 underway: Q-0026 identity repair + Phase 0 contracts **#588**, P1A Access Map projection **#589**, P0C groundwork **#591**, P0C seam conversion + P1B `routing_access_conflict` **#592**; **P1B remainder shipped in #632 (2026-06-09, execution-plan Lane 2 — verify merged on live GitHub):** the Q-0045 governance tier-input path (`GovernanceContext.member_tier`, declared tier preferred verbatim, simulation-labeled per §16.4) + the `help_advertises_locked` drift provider + the full Q-0036 denial-copy **draft** (in the PR body for maintainer read-through — **not live-wired**; wiring follows his markup). **P1C executed 2026-06-10 (consolidated plan Batch 5, #656 — verify merged):** Access Map + Help Preview shipped as **staff-hub subpanels, no new command names** (Q-0032), on the tier path as-is. **Next: P2** Feature Profile preview (own planning first) — and the Batch 6 Help projection seam consumes this lane. Q-0028–Q-0031 + Q-0033 are also **answered** (catalogue committed · availability owns quiet mode · snapshots compound+high-risk · risk policy approved · account links deferred — router §20). Route in: plan §16.8.
